requirements.txt
# Core AI / API
anthropic>=0.40.0

# UI
gradio>=4.20.0

# RAG pipeline
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=4.0.0

# PDF generation (for create_docs.py placeholder notes)
fpdf2>=2.7.0
